LayoutTests/inspector/timeline/timeline-event-TimerFire.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="./resources/timeline-event-utilities.js"></script>
<script>
function testSetTimeout() {
    let setTimeoutIdentifier = setTimeout(() => {
        TestPage.addResult("setTimeout fired");
        finishRecording({setTimeoutIdentifier});
    });
}

function testSetInterval() {
    let count = 0;
    let setIntervalIdentifier = setInterval(() => {
        count++;
        TestPage.addResult("setInterval fired: " + count);
        if (count === 3) {
            clearInterval(setIntervalIdentifier);
            finishRecording({setIntervalIdentifier});
        }
    });
}

function test()
{
    let suite = InspectorTest.createAsyncSuite("TimelineEvent.TimerFire");

    let timeoutIdentifier;
    let intervalIdentifier;

    suite.addTestCase({
        name: "TimelineEvent.TimerFire.setTimeout",
        test(resolve, reject) {
            captureTimelineWithScript(`testSetTimeout()`).then(() => {
                InspectorTest.assert(typeof pageRecordingData.setTimeoutIdentifier === "number");
                timeoutIdentifier = pageRecordingData.setTimeoutIdentifier;

                let recording = WI.timelineManager.activeRecording;
                let scriptTimeline = recording.timelines.get(WI.TimelineRecord.Type.Script);
                let records = scriptTimeline.records.filter((x) => x.eventType === WI.ScriptTimelineRecord.EventType.TimerFired);
                InspectorTest.expectEqual(records.length, 1, "Should be 1 TimerFired record.");
                for (let record of records) {
                    InspectorTest.log("DETAILS: " + typeof record.details);
                    InspectorTest.expectEqual(record.details, timeoutIdentifier, "ScriptTimelineRecord details should be the setTimeout identifier.");
                }
            }).then(resolve, reject);
        }
    });

    suite.addTestCase({
        name: "TimelineEvent.TimerFire.setInterval",
        test(resolve, reject) {
            captureTimelineWithScript(`testSetInterval()`).then(() => {
                InspectorTest.assert(typeof pageRecordingData.setIntervalIdentifier === "number");
                intervalIdentifier = pageRecordingData.setIntervalIdentifier;

                let recording = WI.timelineManager.activeRecording;
                let scriptTimeline = recording.timelines.get(WI.TimelineRecord.Type.Script);
                let records = scriptTimeline.records.filter((x) => x.eventType === WI.ScriptTimelineRecord.EventType.TimerFired);
                InspectorTest.expectEqual(records.length, 3, "Should be 3 TimerFired records.");
                for (let record of records) {
                    InspectorTest.log("DETAILS: " + typeof record.details);
                    InspectorTest.expectEqual(record.details, pageRecordingData.setIntervalIdentifier, "ScriptTimelineRecord details should be the setInterval identifier.");
                }
            }).then(resolve, reject);
        }
    });

    suite.addTestCase({
        name: "SanityCheck",
        test(resolve, reject) {
            InspectorTest.expectThat(timeoutIdentifier !== intervalIdentifier, "The setTimeout identifier and setInterval identifiers should be different.");
            resolve();
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Tests 'TimerFire' Timeline event records.</p>
</body>
</html>
